+++ R/utils.R
@@ -89,8 +89,12 @@ str_remove <- function(x, pattern) {
   str_replace(x, pattern, "")
 }
 
-
 is_tags <- function(x) {
   inherits(x, "shiny.tag") ||
     inherits(x, "shiny.tag.list")
 }
+
+knitr_engine <- function(engine) {
+  engine %||% "r"
+}
+