+++ tests/testthat/test-available-tutorials.R
@@ -25,8 +25,8 @@ test_that("Tutorial names are retrieved", {
   - ex-data-summarise : \"Summarise Tables\"
   - ex-setup-r        : \"Set Up\"
   - hello             : \"Hello, Tutorial!\"
-  - hello-chained     : \"Chained setup chunks\"
   - quiz_question     : \"Tutorial Quiz Questions in `learnr`\"
+  - setup-chunks      : \"Chained setup chunks\"
   - slidy             : \"Slidly demo\""
   )